+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[4.7.0] - 2016-05-04
+---------------------
+##### Added
+- Proper namespace support.
+- Comment widget.
+
 [4.6.0] - 2016-04-26
 ---------------------
 ##### Added